ASS {\move()}: If move distance was negative, the objects were not displayed.
git-svn-id: https://ffdshow-tryout.svn.sourceforge.net/svnroot/ffdshow-tryout/trunk@2693 3b938f2f-1a1a-0410-8054-a526ea5ff92c

+++ src/subtitles/TsubtitleProps.cpp
@@ -359,13 +359,13 @@ int TSubtitleProps::get_maxWidth(unsigned int screenWidth, int subFormat, Iffdsh
 int TSubtitleProps::get_movedistanceV(unsigned int screenHeight) const
 {
  if (!isMove || !refResY) return 0;
- return (posy2-posy)*screenHeight/refResY;
+ return (posy2-posy)*(int)screenHeight/(int)refResY;
 }
 
 int TSubtitleProps::get_movedistanceH(unsigned int screenWidth) const
 {
  if (!isMove || !refResX) return 0;
- return (posx2-posx)*screenWidth/refResX;
+ return (posx2-posx)*(int)screenWidth/(int)refResX;
 }
 
 REFERENCE_TIME TSubtitleProps::get_moveStart() const
